added lights and updated the xmlloader & fileloader
further I deleted the old stuff from the buildingparts & objects and added
all existing objects to the objects.xml. It may be that I falsely
overwrote the files index.html; style.css; firstperson.js and player.css,
it may be that there are mergeconflicts... In which case I think the one
on github is correct (mine is still in english and so on, I think somebody
changed that). ^